+++ src/javascript/languages.js
@@ -17,7 +17,7 @@ let languages = [];
  *
  */
 
-(() => {
+window.onload = function(){
   fs.readdir(path.join(__dirname, "languages"), (err, paths) => {
     paths.forEach(dir => {
       fs.readFile(path.join(__dirname, "languages", dir), "utf8", function (
@@ -30,30 +30,29 @@ let languages = [];
         } catch {
           return;
         }
-        const obj = JSON.parse(data);
-        languages.push(obj); // Push the language
+        languages.push(JSON.parse(data)); // Push the language
         if (languages.length === paths.length) {
-          loadConfig();
+          loadConfig(); 
         }
       });
     });
   });
-})()
+}
 
 const loadLanguage = language => {
   languages.map((item, index) => {
-    if (item["g_l"] === language) {
-      current_config["language"] = item;
+    if (item["name"] === language) {
+      current_config.language = item;
       const toTranslate = document.getElementsByClassName("translate_word");
       for (i = 0; i < toTranslate.length; i++) {
-        if (item[toTranslate[i].getAttribute("idT")] != undefined) {
-          toTranslate[i].innerText = item[toTranslate[i].getAttribute("idT")];
+        if (item.strings[toTranslate[i].getAttribute("idT")] != undefined) {
+          toTranslate[i].innerText = item.strings[toTranslate[i].getAttribute("idT")];
         } else {
           languages.map((item, index) => {
-            if (item["g_l"] === "english") {
-              if (item[toTranslate[i].getAttribute("idT")] != undefined)
+            if (item["name"] === "english") {
+              if (item.strings[toTranslate[i].getAttribute("idT")] != undefined)
                 toTranslate[i].innerText =
-                item[toTranslate[i].getAttribute("idT")];
+                  item.strings[toTranslate[i].getAttribute("idT")];
             }
           });
         }
@@ -61,14 +60,15 @@ const loadLanguage = language => {
     }
   });
 };
+
 const getTranslation = text => {
-  if (current_config.language[text] == undefined) {
+  if (current_config.language.strings[text] == undefined) {
     for (i = 0; i < languages.length; i++) {
-      if (languages[i]["g_l"] === "english") {
-        return languages[i][text] != undefined ? languages[i][text] : text;
+      if (languages[i]["name"] === "english") {
+        return languages[i].strings[text] != undefined ? languages[i].strings[text] : text;
       }
     }
   } else {
-    return current_config.language[text];
+    return current_config.language.strings[text];
   }
 };